requirements.txt
fastapi
uvicorn
duckdb
streamlit
openai
sentence-transformers
pandas
plotly
sqlmodel
python-dotenv
